+++ CHANGES.md
@@ -1,4 +1,5 @@
 Release with following new features and fixes:
 - [WebInterface-OnBoot](https://github.com/rM-self-serve/webinterface-onboot) is currently not supported by reMarkable software version 3.16 or higher
+- Downloading a PDF without synchronization is now possible (context menu)
 - Uploading a template requires a PNG or SVG file, no longer both
 - Build against Avalonia 11.2.3